server/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
billiard==4.2.1
boto3==1.34.131
botocore==1.34.162
celery==5.5.3
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cobble==0.1.4
colorama==0.4.6
cryptography==45.0.4
distro==1.9.0
dj-database-url==2.1.0
Django==5.2.1
django-cors-headers==4.3.1
django-filter==25.1
django-redis==5.3.0
django-storages==1.14.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.0
filetype==1.2.0
grpcio==1.73.1
gunicorn==21.2.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
kombu==5.5.4
libmagic==1.0
lxml==5.4.0
mammoth==1.9.1
numpy==1.24.3
openai==1.88.0
packaging==25.0
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.2.1
portalocker==2.10.1
prompt_toolkit==3.0.51
protobuf==6.31.1
psycopg2-binary==2.9.9
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.9.0
pypdfium2==4.30.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.2.0
python-magic==0.4.27

#pywin32==306 remove for docker builds
qdrant-client==1.14.3
redis==6.2.0
requests==2.31.0
s3transfer==0.10.4
scikit-learn==1.3.0
scipy==1.15.3
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.7.0
